backend/requirements.txt
asgiref==3.8.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.6.0
django-phonenumber-field==8.3.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
idna==3.11
jmespath==1.0.1
mysqlclient==2.2.7
phonenumbers==9.0.16
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3


# Blockchain Integration - Required Python Packages

# AWS SDK
boto3==1.42.0
botocore==1.41.6

# Core blockchain
web3==6.15.0
eth-account==0.11.0
eth-utils==2.3.1

# Solidity compiler (for contract deployment)
py-solc-x==2.0.2

# Async tasks
celery==5.3.6
redis==5.0.1

# Model field tracking
django-model-utils==4.3.1

# For better logging
python-json-logger==2.0.7
